Recursive struct flattening, rename/drop/select API, and codebase cleanup (#52)
* Recursive struct flattening + rename API for columnar serde DataFrame

Nested structs in `vec_to_dataframe` are now recursively flattened into
prefixed columns (e.g., `point_x`, `point_y`). Key features:

- Recursive schema discovery with `FieldMap` for name-based column routing
- `#[serde(flatten)]` support — flattened fields appear without prefix
- `#[serde(skip_serializing_if)]` — skipped fields get NA via padding
- `Option<Struct>` → all sub-columns get NA when None
- `dataframe(&rows).rename("old", "new").build()` builder API

Replaces the old positional filling approach which broke on
skip_serializing_if and couldn't recurse into nested structs.


* Fix pre-existing CI issues: clippy, connection tests, jsonlite

- Move `mod tests;` from io_adapters.rs to connection.rs (wrong module path)
- Replace `b"...\0".as_ptr().cast()` with `c"..."` c-string literals (clippy)
- Fix clippy `let_and_return` in ndarray/nalgebra impls
- Add jsonlite to Suggests (used in tests but undeclared)


* Auto-generate @param for undocumented function parameters

The miniextendr wrapper generator now emits @param entries for all
function parameters that lack Rust doc @param comments. This fixes
R CMD check warnings about undocumented arguments on oldrel-1.

Also regenerates all Rd files and syncs template patch.


* Skip S7 fallback test on R < 4.5 (segfaults on oldrel-1)

The `@` operator on non-S7 objects crashes R 4.4.x before tryCatch
can intercept it. Skip the test on older R versions.


* Fix S7 fallback segfault on R < 4.5: use inherits() instead of @

The generated wrapper for S7 class_any (fallback) methods used
`tryCatch(x@.ptr, error = ...)` to extract the external pointer.
On R 4.4.x, `@` on non-S7 objects segfaults before tryCatch can
catch it. Replace with `if (inherits(x, "S7_object")) x@.ptr else x`
which avoids calling `@` on non-S7 objects entirely.


* Rebuild vendor tarball with S7 fallback fix

The CI builds from inst/vendor.tar.xz in CRAN mode. The previous
tarball still had the old tryCatch(x@.ptr) code that segfaults on
R 4.4. Rebuild to include the inherits() fix.


* Fix S7 fallback: stop() on wrong type instead of passing to Rust

The previous fix (inherits check) avoided the @ segfault but still
passed the raw non-pointer value to .Call, which segfaults in Rust.
Now errors in R with a clear message before reaching Rust.


* Add debug_assert EXTPTRSXP type check in ExternalPtr entry points

In debug builds, from_sexp_unchecked and wrap_sexp now assert that
the SEXP is actually an EXTPTRSXP before dereferencing. Catches
wrong-type-passed-to-Rust bugs early instead of silent UB/segfault.
Zero cost in release builds.


* Replace TYPEOF(x) with x.type_of() across codebase

Use the SexpExt::type_of() method instead of raw ffi::TYPEOF() calls.
Removes unnecessary unsafe blocks and `as SEXPTYPE` casts since
type_of() is safe and already returns SEXPTYPE.

Also adds debug_assert_eq EXTPTRSXP checks to all ExternalPtr entry
points (wrap_sexp_unchecked, wrap_sexp_with_error).


---------

+++ miniextendr-bench/benches/gc_protect.rs
@@ -266,7 +266,7 @@ fn list_builder_construction(size_idx: usize) {
     let n = LIST_SIZES[size_idx];
     unsafe {
         let scope = ProtectScope::new();
-        let builder = ListBuilder::new(&scope, n);
+        let builder = ListBuilder::new(&scope, n as usize);
 
         for i in 0..n {
             let child = scope.protect_raw(ffi::Rf_ScalarInteger(i as i32));
@@ -339,7 +339,7 @@ fn strvec_builder_construction(size_idx: usize) {
     let n = LIST_SIZES[size_idx];
     unsafe {
         let scope = ProtectScope::new();
-        let builder = StrVecBuilder::new(&scope, n);
+        let builder = StrVecBuilder::new(&scope, n as usize);
 
         for i in 0..n {
             builder.set_str(i, "hello");
@@ -502,7 +502,7 @@ fn strvec_set_str_high_count(n: usize) {
 fn nested_lists_reprotect(outer_count: usize) {
     unsafe {
         let scope = ProtectScope::new();
-        let outer = ListBuilder::new(&scope, outer_count as isize);
+        let outer = ListBuilder::new(&scope, outer_count);
 
         for i in 0..outer_count {
             // Use reprotect slot for inner list - each iteration reuses slot
@@ -526,7 +526,7 @@ fn nested_lists_reprotect(outer_count: usize) {
 fn list_builder_grows_stack(n: usize) {
     unsafe {
         let scope = ProtectScope::new();
-        let builder = ListBuilder::new(&scope, n as isize);
+        let builder = ListBuilder::new(&scope, n);
 
         for i in 0..n {
             // Each child adds to protect stack
@@ -679,7 +679,7 @@ fn collect_into_vec_then_list(n: usize) {
         let values: Vec<i32> = (0..n).map(|i| i as i32).collect();
 
         // Then build list with known size
-        let builder = ListBuilder::new(&scope, values.len() as isize);
+        let builder = ListBuilder::new(&scope, values.len());
         for (i, v) in values.into_iter().enumerate() {
             builder.set_protected(i as isize, ffi::Rf_ScalarInteger(v));
         }
@@ -695,7 +695,7 @@ fn accumulator_vs_builder_known_size(n: usize) {
     unsafe {
         // ListBuilder (optimal for known size)
         let scope1 = ProtectScope::new();
-        let builder = ListBuilder::new(&scope1, n as isize);
+        let builder = ListBuilder::new(&scope1, n);
         for i in 0..n {
             builder.set_protected(i as isize, ffi::Rf_ScalarInteger(i as i32));
         }
